backup-reorganized/active-prototypes/analytics/pyproject.toml
[project]
name = "a2a-samples-analytics"
version = "0.1.0"
description = "Generate charts from CSV-like data passed in"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk[sqlite]>=0.2.6",
    "click>=8.1.8",
    "crewai[tools]>=0.95.0",
    "matplotlib>=3.8.0",
    "pandas>=2.1.0",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
]
